+++ src/utils/interp.rs
@@ -1,9 +1,9 @@
 pub struct Akima {
     len: usize,
-    coeffs: Vec<[f64; 4]>,
+    coeffs: Vec<[f32; 4]>,
 }
 impl Akima {
-    pub fn new(y: &[f64]) -> Self {
+    pub fn new(y: &[f32]) -> Self {
         let n = y.len();
         let mut m = Vec::with_capacity(n + 3);
         m.push(0.0);
@@ -35,15 +35,15 @@ impl Akima {
         Self { len: n, coeffs }
     }
     #[inline(always)]
-    pub fn sample_with_slice(&self, x: &[f64]) -> Vec<f64> {
+    pub fn sample_with_slice(&self, x: &[f32]) -> Vec<f32> {
         let mut res = Vec::with_capacity(x.len());
         let mut i = 0;
         for &t in x {
-            while i < self.len - 1 && ((i + 1) as f64) < t {
+            while i < self.len - 1 && ((i + 1) as f32) < t {
                 i += 1;
             }
             let [c0, c1, c2, c3] = self.coeffs[i];
-            let r = t - i as f64;
+            let r = t - i as f32;
             res.push(c0 + r * (c1 + r * (c2 + r * c3)));
         }
         res
