+++ src/utils/stft.rs
@@ -3,13 +3,13 @@ use ndarray::{Array2, ArrayView1, s};
 use once_cell::sync::{Lazy, OnceCell};
 use dashmap::DashMap;
 use oxifft::{Complex, Direction, Flags, Plan, streaming::WindowFunction, threading::{get_default_pool, ThreadPool}};
-static HANN_WINDOWS: Lazy<DashMap<usize, Arc<Vec<f64>>>> = Lazy::new(DashMap::new);
-static FFT_PLANS: Lazy<DashMap<(usize, Direction), Arc<Plan<f64>>>> = Lazy::new(DashMap::new);
-static ISTFT_WINDOW_SQ: Lazy<Arc<Vec<f64>>> = Lazy::new(|| {
+static HANN_WINDOWS: Lazy<DashMap<usize, Arc<Vec<f32>>>> = Lazy::new(DashMap::new);
+static FFT_PLANS: Lazy<DashMap<(usize, Direction), Arc<Plan<f32>>>> = Lazy::new(DashMap::new);
+static ISTFT_WINDOW_SQ: Lazy<Arc<Vec<f32>>> = Lazy::new(|| {
     let window = get_hann_window(crate::consts::FFT_SIZE);
     Arc::new(window.iter().map(|&w| w * w).collect())
 });
-fn get_hann_window(fft_size: usize) -> Arc<Vec<f64>> {
+fn get_hann_window(fft_size: usize) -> Arc<Vec<f32>> {
     HANN_WINDOWS
         .entry(fft_size)
         .or_insert_with(|| {
@@ -17,7 +17,7 @@ fn get_hann_window(fft_size: usize) -> Arc<Vec<f64>> {
         })
         .clone()
 }
-fn get_fft_plan(fft_size: usize, direction: Direction) -> Arc<Plan<f64>> {
+fn get_fft_plan(fft_size: usize, direction: Direction) -> Arc<Plan<f32>> {
     FFT_PLANS
         .entry((fft_size, direction))
         .or_insert_with(|| {
@@ -29,10 +29,10 @@ fn get_fft_plan(fft_size: usize, direction: Direction) -> Arc<Plan<f64>> {
         .clone()
 }
 pub fn stft_core(
-    signal: &[f64],
+    signal: &[f32],
     fft_size: usize,
     hop_size: usize,
-) -> Array2<Complex<f64>> {
+) -> Array2<Complex<f32>> {
     let freq_bins = fft_size / 2 + 1; 
     if fft_size == 0 || hop_size == 0 || signal.len() < fft_size {
         return Array2::from_shape_vec((freq_bins, 0), Vec::new()).unwrap();
@@ -42,14 +42,14 @@ pub fn stft_core(
     let n_frames = (signal.len() - fft_size) / hop_size + 1;
     let mut spec = Array2::from_shape_fn((freq_bins, n_frames), |_| Complex::zero()); 
     let pool = get_default_pool();
-    let result: Arc<Vec<OnceCell<Vec<Complex<f64>>>>> = Arc::new(
+    let result: Arc<Vec<OnceCell<Vec<Complex<f32>>>>> = Arc::new(
         (0..n_frames)
             .map(|_| OnceCell::new()) 
             .collect()
     );
     pool.parallel_for(n_frames, |frame_idx| {
         let start = frame_idx * hop_size;
-        let input: Vec<Complex<f64>> = signal[start..start + fft_size]
+        let input: Vec<Complex<f32>> = signal[start..start + fft_size]
             .iter()
             .zip(window.iter())
             .map(|(&s, &w)| Complex::new(s * w, 0.0))
@@ -65,11 +65,11 @@ pub fn stft_core(
     spec
 }
 pub fn istft_core(
-    spec: &Array2<Complex<f64>>,
+    spec: &Array2<Complex<f32>>,
     target_len: usize,
     fft_size: usize,
     hop_size: usize,
-) -> Vec<f64> {
+) -> Vec<f32> {
     let (freq_bins, n_frames) = (spec.nrows(), spec.ncols());
     if n_frames == 0 || freq_bins == 0 || freq_bins != fft_size / 2 + 1 {
         return vec![0.0; target_len];
@@ -79,9 +79,9 @@ pub fn istft_core(
     let out_len = fft_size + (n_frames - 1) * hop_size;
     let mut output = vec![0.0; out_len];
     let mut win_sum = vec![0.0; out_len];
-    let scale = 1.0 / fft_size as f64;
+    let scale = 1.0 / fft_size as f32;
     let pool = get_default_pool();
-    let result: Arc<Vec<OnceCell<Vec<f64>>>> = Arc::new(
+    let result: Arc<Vec<OnceCell<Vec<f32>>>> = Arc::new(
         (0..n_frames)
             .map(|_| OnceCell::new())
             .collect()
@@ -99,7 +99,7 @@ pub fn istft_core(
             full_spec[fft_size - i] = full_spec[i].conj();
         }
         plan.execute(&full_spec, &mut frame);
-        let ifft_result: Vec<f64> = frame
+        let ifft_result: Vec<f32> = frame
             .iter()
             .zip(window.iter())
             .map(|(frame_val, win_val)| frame_val.re * scale * win_val)
